Fix conversion for create trigger RS-194164

+++ ora2rdb/src/main/java/ru/redsoft/ora2rdb/RewritingListener.java
@@ -2831,10 +2831,6 @@ public class RewritingListener extends PlSqlParserBaseListener {
         insertBefore(ctx.trigger_body(), indentation + position + " SQL SECURITY DEFINER\nAS\n");
 
         if (ctx.simple_dml_trigger() != null) {
-            if (ctx.simple_dml_trigger().for_each_row() == null) {
-                replace(ctx, " -unconvertible  RS-228329 \n" + getRewriterText(ctx));
-                current_plsql_block.commentBlock = true;
-            }
             if (ctx.simple_dml_trigger().DISABLE() != null)
                 insertAfter(ctx, "\nALTER TRIGGER " + Ora2rdb.getRealName(ctx.trigger_name().getText()) + " INACTIVE;");
 
@@ -2843,10 +2839,6 @@ public class RewritingListener extends PlSqlParserBaseListener {
         }
 
         if (ctx.compound_dml_trigger() != null) {
-            replace(ctx.compound_dml_trigger(), "[-unconvertible RS-228297 " + getRewriterText(ctx.compound_dml_trigger()));
-            insertAfter(ctx.trigger_body().compound_trigger_block(), "]");
-            current_plsql_block.commentBlock = true;
-
             if (ctx.compound_dml_trigger().DISABLE() != null)
                 insertAfter(ctx, "\nALTER TRIGGER " + Ora2rdb.getRealName(ctx.trigger_name().getText()) + " INACTIVE;");
 
@@ -2860,50 +2852,36 @@ public class RewritingListener extends PlSqlParserBaseListener {
 
             delete(ctx.non_dml_trigger().DISABLE());
             delete(ctx.non_dml_trigger().ENABLE());
-
-            if (ctx.non_dml_trigger().INSTEAD() != null) {
-                insertBefore(ctx.non_dml_trigger(), "[-unconvertible RS-228348 ");
-                insertAfter(ctx.non_dml_trigger(), "]");
-                current_plsql_block.commentBlock = true;
-            } else {
-                for (Non_dml_eventContext stmt : ctx.non_dml_trigger().non_dml_event()) {
-                    if (stmt.database_event() != null) {
-                        if (stmt.database_event().LOGON() != null) {
-                            replace(stmt.database_event().LOGON(), "ON CONNECT");
-                        } else if (stmt.database_event().LOGOFF() != null) {
-                            replace(stmt.database_event().LOGOFF(), "ON DISCONNECT");
-                        } else {
-                            insertBefore(stmt.database_event(), "[-unconvertible RS-228336 ");
-                            insertAfter(stmt.database_event(), "]");
-                            current_plsql_block.commentBlock = true;
-                        }
-                        delete(ctx.non_dml_trigger().AFTER());
-                        delete(ctx.non_dml_trigger().BEFORE());
+            for (Non_dml_eventContext stmt : ctx.non_dml_trigger().non_dml_event()) {
+                if (stmt.database_event() != null) {
+                    if (stmt.database_event().LOGON() != null) {
+                        replace(stmt.database_event().LOGON(), "ON CONNECT");
+                    } else if (stmt.database_event().LOGOFF() != null) {
+                        replace(stmt.database_event().LOGOFF(), "ON DISCONNECT");
                     }
-                    if (stmt.ddl_event() != null) {
-                        if (stmt.ddl_event().CREATE() != null || stmt.ddl_event().ALTER() != null
-                                || stmt.ddl_event().DROP() != null || stmt.ddl_event().DDL() != null) {
-                            if (ctx.non_dml_trigger().AFTER() != null)
-                                replace(ctx.non_dml_trigger(), " AFTER ANY DDL STATEMENT");
-                            else if (ctx.non_dml_trigger().BEFORE() != null)
-                                replace(ctx.non_dml_trigger(), " BEFORE ANY DDL STATEMENT");
-                            return;
-                        } else {
-                            insertBefore(stmt.ddl_event(), "[-unconvertible RS-228339 ");
-                            insertAfter(stmt.ddl_event(), "]");
-                            current_plsql_block.commentBlock = true;
-                        }
+                    delete(ctx.non_dml_trigger().AFTER());
+                    delete(ctx.non_dml_trigger().BEFORE());
+                }
+                if (stmt.ddl_event() != null) {
+                    if (stmt.ddl_event().CREATE() != null || stmt.ddl_event().ALTER() != null
+                            || stmt.ddl_event().DROP() != null || stmt.ddl_event().DDL() != null) {
+                        if (ctx.non_dml_trigger().AFTER() != null)
+                            replace(ctx.non_dml_trigger(), " AFTER ANY DDL STATEMENT");
+                        else if (ctx.non_dml_trigger().BEFORE() != null)
+                            replace(ctx.non_dml_trigger(), " BEFORE ANY DDL STATEMENT");
+                        return;
                     }
                 }
             }
+
             delete(ctx.non_dml_trigger().ON());
             delete(ctx.non_dml_trigger().SCHEMA());
             delete(ctx.non_dml_trigger().schema_name());
             delete(ctx.non_dml_trigger().PLUGGABLE());
             delete(ctx.non_dml_trigger().DATABASE());
         }
-
         if (!Ora2rdb.reorder)
+
             insertBefore(ctx, temp_tables_ddl + "\n");
 //            replace(ctx, temp_tables_ddl + getRewriterText(ctx));
         else
@@ -2924,6 +2902,7 @@ public class RewritingListener extends PlSqlParserBaseListener {
 
     @Override
     public void exitReferencing_clause(Referencing_clauseContext ctx) {
+        boolean deleteCtx = true;
         if (current_plsql_block != null) {
             for (int i = 0; i < ctx.referencing_element().size(); i++) {
                 Referencing_elementContext stmt = ctx.referencing_element().get(i);
@@ -2933,16 +2912,15 @@ public class RewritingListener extends PlSqlParserBaseListener {
                 if (stmt.OLD() != null) {
                     current_plsql_block.trigger_referencing_attributes.oldValue = Ora2rdb.getRealName(stmt.column_alias().identifier().getText());
                 }
-                if (stmt.PARENT() != null) {
-                    replace(stmt.PARENT(), "[-unconvertible RS-228325 " + stmt.PARENT());
-                    replace(stmt.column_alias(), getRewriterText(stmt.column_alias()) + "]");
-                    current_plsql_block.commentBlock = true;
-                    return;
-                }
+                if (stmt.PARENT() != null)
+                    deleteCtx = false;
             }
+            if (deleteCtx){
+                delete(ctx);
+                deleteSPACESLeft(ctx);
+            }
+
         }
-        delete(ctx);
-        deleteSPACESLeft(ctx);
     }
 
     @Override
@@ -2995,13 +2973,6 @@ public class RewritingListener extends PlSqlParserBaseListener {
         if (ctx.column_list() != null) {
             delete(ctx.column_list().column_name());
         }
-    }
-
-    @Override
-    public void exitDml_event_nested_clause(Dml_event_nested_clauseContext ctx) {
-        replace(ctx, "[-unconvertible RS-228312 " + getRewriterText(ctx) + "]");
-        if (current_plsql_block != null)
-            current_plsql_block.commentBlock = true;
     }
 
     @Override
